requirements.txt
requests
beautifulsoup4
whois
PyPDF2
python-docx
colorama
Flask
jinja2
aiohttp
dnspython
